Population Data
Got data from US census API, and map on a choropleth based of the estimate population density in the block

TODO: Blocks are large polygons, and currently I have intersections. Need to multiply by percentage of intersect to get a new estimated number of coverage based of the distance that can be reached.